+++ EngineLib/Renderer/Texture.c
@@ -48,43 +48,43 @@ Macro definitions
 Functions declarations
 ************************************************************************************************************************/
 
-TextureResultType Texture_Init( Texture* texture, uint32_t width, uint32_t height, uint32_t channels )
+TextureResultType Texture_Init(Texture* texture, uint32_t width, uint32_t height, uint32_t channels)
 {
-    texture->info.width = width;
-    texture->info.height = height;
+    texture->info.width = (float) width;
+    texture->info.height = (float) height;
     texture->info.channels = channels;
     texture->data = NULL;
 
-    if ( texture->info.width == 0 || texture->info.height == 0 || texture->info.channels == 0 )
+    if (texture->info.width == 0 || texture->info.height == 0 || texture->info.channels == 0)
     {
         return TextureResult_Error;
     }
     SDL_PixelFormat format;
-    if ( channels >= 4 ) { format = SDL_PIXELFORMAT_RGBA8888; }
-    else if ( channels == 3 ) { format = SDL_PIXELFORMAT_XRGB8888; }
-    else if ( channels == 2 ) { format = SDL_PIXELFORMAT_RGB565; }
+    if (channels >= 4) { format = SDL_PIXELFORMAT_RGBA8888; }
+    else if (channels == 3) { format = SDL_PIXELFORMAT_XRGB8888; }
+    else if (channels == 2) { format = SDL_PIXELFORMAT_RGB565; }
     else { return TextureResult_Error; }
 
-    texture->data =
-            ( void* ) SDL_CreateTexture( RendererGet()->renderer, format, SDL_TEXTUREACCESS_STATIC, width, height );
+    texture->data = (void*) SDL_CreateTexture(RendererGet()->renderer, format, SDL_TEXTUREACCESS_STATIC, (int) width,
+                                              (int) height);
     return TextureResult_Success;
 }
 
-TextureResultType Texture_SetData( Texture* texture, void* data )
+TextureResultType Texture_SetData(Texture* texture, void* data)
 {
     texture->data = data;
     return TextureResult_Success;
 }
 
-TextureResultType Texture_GetInfo( Texture* texture, TextureInfo** info )
+TextureResultType Texture_GetInfo(Texture* texture, TextureInfo** info)
 {
     *info = &texture->info;
     return TextureResult_Success;
 }
 
-TextureResultType Texture_Destroy( Texture* texture )
+TextureResultType Texture_Destroy(Texture* texture)
 {
-    SDL_DestroyTexture( texture->data );
+    SDL_DestroyTexture(texture->data);
     texture->data = NULL;
     return TextureResult_Success;
 }
