+++ EngineLib/Renderer/Texture.h
@@ -51,8 +51,8 @@ Macro definitions
 Type definitions
 ***********************************************************************************************************************/
 typedef struct {
-    uint32_t width;
-    uint32_t height;
+    float width;
+    float height;
     uint32_t channels;
 } TextureInfo;
